// supabase/types.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      files: {
        Row: {
          chat_id: string | null;
          created_at: string;
          file_path: string;
          id: string;
          message_id: string | null;
          name: string;
          sharing: string;
          size: number;
          tokens: number;
          type: string;
          updated_at: string | null;
          user_id: string;
        };
        Insert: {
          chat_id?: string | null;
          created_at?: string;
          file_path: string;
          id?: string;
          message_id?: string | null;
          name: string;
          sharing?: string;
          size: number;
          tokens: number;
          type: string;
          updated_at?: string | null;
          user_id: string;
        };
        Update: {
          chat_id?: string | null;
          created_at?: string;
          file_path?: string;
          id?: string;
          message_id?: string | null;
          name?: string;
          sharing?: string;
          size?: number;
          tokens?: number;
          type?: string;
          updated_at?: string | null;
          user_id?: string;
        };
        Relationships: [
          {
            foreignKeyName: 'files_chat_id_fkey';
            columns: ['chat_id'];
            isOneToOne: false;
            referencedRelation: 'chats';
            referencedColumns: ['id'];
          },
          {
            foreignKeyName: 'files_message_id_fkey';
            columns: ['message_id'];
            isOneToOne: false;
            referencedRelation: 'messages';
            referencedColumns: ['id'];
          },
        ];
      };
// ...
    };
// ...
  };
// ...
};

type DefaultSchema = Database[Extract<keyof Database, 'public'>];
// ...
export type TablesInsert<
  DefaultSchemaTableNameOrOptions extends
    | keyof DefaultSchema['Tables']
    | { schema: keyof Database },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof Database;
  }
    ? keyof Database[DefaultSchemaTableNameOrOptions['schema']]['Tables']
    : never = never,
> = DefaultSchemaTableNameOrOptions extends { schema: keyof Database }
  ? Database[DefaultSchemaTableNameOrOptions['schema']]['Tables'][TableName] extends {
      Insert: infer I;
    }
    ? I
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema['Tables']
    ? DefaultSchema['Tables'][DefaultSchemaTableNameOrOptions] extends {
        Insert: infer I;
      }
      ? I
      : never
    : never;